* commit '435957670b45035ccc9cd1e50c3e824f4a1dd4d6':
  Added dropping and creating v_marker_linkage_physical to be able to change column type.
  Altered columnn data type for linkage group start and stop.

+++ builder/liquibase/changelogs/changesets/run_on_change/create_functions.sql
@@ -178,7 +178,7 @@ CREATE OR REPLACE FUNCTION createjob(_name text, _type text, _payload_type text,
     END;
 $$;
 
-CREATE OR REPLACE FUNCTION createlinkagegroup(linkagegroupname text, linkagegroupstart integer, linkagegroupstop integer, mapid integer, createdby integer, createddate date, modifiedby integer, modifieddate date, OUT id integer) RETURNS integer
+CREATE OR REPLACE FUNCTION createlinkagegroup(linkagegroupname text, linkagegroupstart numeric, linkagegroupstop numeric, mapid integer, createdby integer, createddate date, modifiedby integer, modifieddate date, OUT id integer) RETURNS integer
     LANGUAGE plpgsql
     AS $$
 	BEGIN
